+++ config/vim/pack/plugins/start/fzf/src/options.go
@@ -127,6 +127,7 @@ Usage: fzf [options]
                              (default: 0 or center)
 
   INPUT SECTION
+    --no-input               Disable and hide the input section
     --prompt=STR             Input prompt (default: '> ')
     --info=STYLE             Finder info style
                              [default|right|hidden|inline[-right][:PREFIX]]
@@ -538,6 +539,7 @@ type Options struct {
 	Scheme            string
 	Extended          bool
 	Phony             bool
+	Inputless         bool
 	Case              Case
 	Normalize         bool
 	Nth               []Range
@@ -659,6 +661,7 @@ func defaultOptions() *Options {
 		Scheme:       "", // Unknown
 		Extended:     true,
 		Phony:        false,
+		Inputless:    false,
 		Case:         CaseSmart,
 		Normalize:    true,
 		Nth:          make([]Range, 0),
@@ -882,7 +885,7 @@ func parseKeyChordsImpl(str string, message string) (map[tui.Event]string, error
 		case "right":
 			add(tui.Right)
 		case "enter", "return":
-			add(tui.CtrlM)
+			add(tui.Enter)
 		case "space":
 			chords[tui.Key(' ')] = key
 		case "backspace", "bspace", "bs":
@@ -1333,7 +1336,7 @@ const (
 
 func init() {
 	executeRegexp = regexp.MustCompile(
-		`(?si)[:+](become|execute(?:-multi|-silent)?|reload(?:-sync)?|preview|(?:change|transform)-(?:query|prompt|(?:border|list|preview|input|header)-label|header|search|nth)|transform|change-(?:preview-window|preview|multi)|(?:re|un)bind|pos|put|print|search)`)
+		`(?si)[:+](become|execute(?:-multi|-silent)?|reload(?:-sync)?|preview|(?:change|transform)-(?:query|prompt|(?:border|list|preview|input|header)-label|header|search|nth)|transform|change-(?:preview-window|preview|multi)|(?:re|un|toggle-)bind|pos|put|print|search)`)
 	splitRegexp = regexp.MustCompile("[,:]+")
 	actionNameRegexp = regexp.MustCompile("(?i)^[a-z-]+")
 }
@@ -1497,6 +1500,12 @@ func parseActionList(masked string, original string, prevActions []*action, putA
 			appendAction(actToggleTrack)
 		case "toggle-track-current":
 			appendAction(actToggleTrackCurrent)
+		case "toggle-input":
+			appendAction(actToggleInput)
+		case "hide-input":
+			appendAction(actHideInput)
+		case "show-input":
+			appendAction(actShowInput)
 		case "toggle-header":
 			appendAction(actToggleHeader)
 		case "toggle-wrap":
@@ -1617,7 +1626,7 @@ func parseActionList(masked string, original string, prevActions []*action, putA
 					actions = append(actions, &action{t: t, a: actionArg})
 				}
 				switch t {
-				case actUnbind, actRebind:
+				case actUnbind, actRebind, actToggleBind:
 					if _, err := parseKeyChordsImpl(actionArg, spec[0:offset]+" target required"); err != nil {
 						return nil, err
 					}
@@ -1702,6 +1711,8 @@ func isExecuteAction(str string) actionType {
 		return actUnbind
 	case "rebind":
 		return actRebind
+	case "toggle-bind":
+		return actToggleBind
 	case "preview":
 		return actPreview
 	case "change-header":
@@ -2315,6 +2326,8 @@ func parseOptions(index *int, opts *Options, allArgs []string) error {
 			opts.Phony = false
 		case "--disabled", "--phony":
 			opts.Phony = true
+		case "--no-input":
+			opts.Inputless = true
 		case "--tiebreak":
 			str, err := nextString("sort criterion required")
 			if err != nil {
@@ -3064,6 +3077,9 @@ func noSeparatorLine(style infoStyle, separator bool) bool {
 }
 
 func (opts *Options) noSeparatorLine() bool {
+	if opts.Inputless {
+		return true
+	}
 	sep := opts.Separator == nil && !opts.InputBorderShape.Visible() || opts.Separator != nil && len(*opts.Separator) > 0
 	return noSeparatorLine(opts.InfoStyle, sep)
 }
@@ -3094,7 +3110,14 @@ func postProcessOptions(opts *Options) error {
 	if opts.HeaderLinesShape == tui.BorderNone {
 		opts.HeaderLinesShape = tui.BorderPhantom
 	} else if opts.HeaderLinesShape == tui.BorderUndefined {
-		opts.HeaderLinesShape = tui.BorderNone
+		// In reverse-list layout, header lines should be at the top, while
+		// ordinary header should be at the bottom. So let's use a separate
+		// window for the header lines.
+		if opts.Layout == layoutReverseList {
+			opts.HeaderLinesShape = tui.BorderPhantom
+		} else {
+			opts.HeaderLinesShape = tui.BorderNone
+		}
 	}
 
 	if opts.Pointer == nil {
@@ -3197,7 +3220,7 @@ func postProcessOptions(opts *Options) error {
 
 	// If 'double-click' is left unbound, bind it to the action bound to 'enter'
 	if _, prs := opts.Keymap[tui.DoubleClick.AsEvent()]; !prs {
-		opts.Keymap[tui.DoubleClick.AsEvent()] = opts.Keymap[tui.CtrlM.AsEvent()]
+		opts.Keymap[tui.DoubleClick.AsEvent()] = opts.Keymap[tui.Enter.AsEvent()]
 	}
 
 	// If we're not using extended search mode, --nth option becomes irrelevant
@@ -3235,7 +3258,13 @@ func postProcessOptions(opts *Options) error {
 
 	// Sets --min-height automatically
 	if opts.Height.size > 0 && opts.Height.percent && opts.MinHeight < 0 {
-		opts.MinHeight = -opts.MinHeight + 1 + borderLines(opts.BorderShape) + borderLines(opts.ListBorderShape) + borderLines(opts.InputBorderShape)
+		opts.MinHeight = -opts.MinHeight + borderLines(opts.BorderShape) + borderLines(opts.ListBorderShape)
+		if !opts.Inputless {
+			opts.MinHeight += 1 + borderLines(opts.InputBorderShape)
+			if !opts.noSeparatorLine() {
+				opts.MinHeight++
+			}
+		}
 		if len(opts.Header) > 0 {
 			opts.MinHeight += borderLines(opts.HeaderBorderShape) + len(opts.Header)
 		}
@@ -3245,9 +3274,6 @@ func postProcessOptions(opts *Options) error {
 				borderShape = opts.HeaderLinesShape
 			}
 			opts.MinHeight += borderLines(borderShape) + opts.HeaderLines
-		}
-		if !opts.noSeparatorLine() {
-			opts.MinHeight++
 		}
 		if len(opts.Preview.command) > 0 && (opts.Preview.position == posUp || opts.Preview.position == posDown) && opts.Preview.Visible() && opts.Preview.position == posUp {
 			borderShape := opts.Preview.border