+++ config/vim/pack/plugins/start/fzf/src/tui/tcell.go
@@ -52,6 +52,7 @@ type TcellWindow struct {
 	borderStyle BorderStyle
 	uri         *string
 	params      *string
+	showCursor  bool
 }
 
 func (w *TcellWindow) Top() int {
@@ -72,7 +73,9 @@ func (w *TcellWindow) Height() int {
 
 func (w *TcellWindow) Refresh() {
 	if w.moveCursor {
-		_screen.ShowCursor(w.left+w.lastX, w.top+w.lastY)
+		if w.showCursor {
+			_screen.ShowCursor(w.left+w.lastX, w.top+w.lastY)
+		}
 		w.moveCursor = false
 	}
 	w.lastX = 0
@@ -102,6 +105,14 @@ const (
 
 func (r *FullscreenRenderer) Bell() {
 	_screen.Beep()
+}
+
+func (r *FullscreenRenderer) HideCursor() {
+	r.showCursor = false
+}
+
+func (r *FullscreenRenderer) ShowCursor() {
+	r.showCursor = true
 }
 
 func (r *FullscreenRenderer) PassThrough(str string) {
@@ -167,6 +178,9 @@ func (r *FullscreenRenderer) getScreen() (tcell.Screen, error) {
 		s, e := tcell.NewScreen()
 		if e != nil {
 			return nil, e
+		}
+		if !r.showCursor {
+			s.HideCursor()
 		}
 		_screen = s
 	}
@@ -546,7 +560,7 @@ func (r *FullscreenRenderer) GetChar() Event {
 
 func (r *FullscreenRenderer) Pause(clear bool) {
 	if clear {
-		_screen.Fini()
+		r.Close()
 	}
 }
 
@@ -558,6 +572,7 @@ func (r *FullscreenRenderer) Resume(clear bool, sigcont bool) {
 
 func (r *FullscreenRenderer) Close() {
 	_screen.Fini()
+	_screen = nil
 }
 
 func (r *FullscreenRenderer) RefreshWindows(windows []Window) {
@@ -590,7 +605,8 @@ func (r *FullscreenRenderer) NewWindow(top int, left int, width int, height int,
 		width:       width,
 		height:      height,
 		normal:      normal,
-		borderStyle: borderStyle}
+		borderStyle: borderStyle,
+		showCursor:  r.showCursor}
 	w.Erase()
 	return w
 }